[Editor] Fixed rounding of Decimal(float) constructor and warnings in VS2010.
Decimal::Decimal(float value) constructor implementation removed from header, value fractional part is now symmetrically rounded.
Return values from methods returning float in Decimal and Vec3 has been promoted to double.
Debug version of _sptmath module is now exported as _sptmathd.

+++ applications/sptEditor/src/sptmath/Vec3.cpp
@@ -19,7 +19,7 @@ Vec3 Vec3::normalized()
     return Vec3(Decimal(x / len), Decimal(y / len), Decimal(z / len));
 };
 
-float Vec3::angleToJUnit() const
+double Vec3::angleToJUnit() const
 {
     double div = double(_y) / length();
     double theta = div <= -1.0 ? double(M_PI) : acos(div);
